requirements2.txt
streamlit
pdf2image
python-dotenv
langchain-groq
PyPDF2
